Final cleanup: remove all remaining Python infrastructure
- Remove pyproject.toml, pytest.ini, requirements.txt, main.py, passenger_wsgi.py
- Remove Python scripts (build_rust, generate_proto, rustify)
- Remove old rust_extensions/axiom_core PyO3 module
- Remove Python API errors and responses modules
- Remove encoding/proto_utils and all generated protobuf Python stubs
- Add build.rs and gRPC module for Rust-native protobuf support
- Update Cargo.toml for final Rust build configuration
- Polish existing Rust modules for cross-module consistency

+++ src/middleware/auth.rs
@@ -39,7 +39,7 @@ pub async fn auth_middleware(
                         db_scope: key_cfg.db_scope.clone(),
                         fs_scope: key_cfg.fs_scope.clone(),
                         feature_scope: key_cfg.feature_scope.clone(),
-                        rate_limit_override: key_cfg.rate_limit_override as u64,
+                        rate_limit_override: key_cfg.rate_limit_override as u32,
                         full_admin: key_cfg.full_admin,
                     };
                     req.extensions_mut().insert(ctx);
